+++ code/src/cljs/sixsq/nuvla/ui/dashboard/utils.cljs
@@ -53,7 +53,7 @@
     (->> deployment-parameters
          (filter is-replicas-running?)
          (map is-value-positive?)
-         (every? true?))                                    ;; careful, this returns true for an empty collection!
+         (every? true?))                      ;; careful, this returns true for an empty collection!
     false))
 
 
@@ -76,3 +76,15 @@
 (defn detail-href
   [id]
   (str "dashboard/" (general-utils/id->uuid id)))
+
+
+(defn get-query-params
+  [full-text-search active-only? page elements-per-page]
+  (let [filter-active-only? (when active-only? "state!='STOPPED'")
+        full-text-search    (when-not (str/blank? full-text-search)
+                              (str "fulltext=='" full-text-search "*'"))
+        filter              (str/join " and " (remove nil? [filter-active-only? full-text-search]))]
+    (cond-> {:first   (inc (* (dec page) elements-per-page))
+             :last    (* page elements-per-page)
+             :orderby "created:desc"}
+            (not (str/blank? filter)) (assoc :filter filter))))